Fix Flaky docker tests
Fixes #27

It seems that the issue was caused by an overloaded file system: Copying
large amounts of data inside the docker containers would cause spikes in
IO usage which, in turn, caused freezes, which then caused test cases to
fail because we're using timeouts in Flexo.

Storing the packages in a tmpfs means that the files are stored in
RAM, not on disk. This is more of a workaround and it also has the
added disadvantage that running the test cases using docker now requires
substantial amounts of RAM. The test cases will most likely not run
successfully with less than 32 GB of RAM.